pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reeln-plugin-google"
dynamic = ["version"]
description = "reeln-cli plugin for Google platform integration (YouTube livestreams, uploads, playlists)"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "python-dateutil",
]

[project.urls]
Homepage = "https://streamn.dad"
Repository = "https://github.com/StreamnDad/reeln-plugin-google"
Changelog = "https://github.com/StreamnDad/reeln-plugin-google/blob/main/CHANGELOG.md"
Issues = "https://github.com/StreamnDad/reeln-plugin-google/issues"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-xdist",
    "ruff",
    "mypy",
]

[project.entry-points."reeln.plugins"]
google = "reeln_google_plugin:GooglePlugin"

[tool.hatch.build.targets.wheel]
packages = ["reeln_google_plugin"]

[tool.hatch.version]
path = "reeln_google_plugin/__init__.py"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "reeln.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "googleapiclient.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google_auth_oauthlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dateutil.*"
ignore_missing_imports = true
